move the constants in staging/src/k8s.io/api/apps to const.go
